requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.1
requests==2.23.0
urllib3==1.25.9
numpy==1.19.5
pillow==8.1.0
pyzbar==0.1.8
qrcode==6.1.0